+++ src/main.js
@@ -1,3 +1,4 @@
+'use strict';
 require('./resizer');
-require('./upload');
-require('./pictures');
+require('./upload/upload');
+require('./pictures/pictures');
